hyperstreamdb::core::index::hnsw_rs::dist

Trait Distance

pub trait Distance<T: Send + Sync> {
    // Required method
    fn eval(&self, va: &[T], vb: &[T]) -> f32;
}

This is the basic Trait describing a distance. The structure Hnsw can be instantiated by anything satisfying this Trait. The crate provides implmentations for L1, L2 , Cosine, Jaccard, Hamming. For other distances implement the trait possibly with the newtype pattern

Required Methods

fn eval(&self, va: &[T], vb: &[T]) -> f32