hyperstreamdb::core::index::hnsw_rs

Module flatten

This module provides conversion of a Point structure to a FlatPoint containing just the Id of a point and those of its neighbours. The whole Hnsw structure is then flattened into a Hashtable associating the data ID of a point to its corresponding FlatPoint.
It can be used, for example, when reloading only the graph part of the data to have knowledge of relative proximity of points as described just by their DataId

Structs

FlatNeighborhood

A structure providing neighbourhood information of a point stored in the Hnsw structure given its DataId.
The structure uses the FlatPoint structure.
This structure can be obtained by FlatNeighborhood::from<&Hnsw<T,D>>

FlatPoint

a reduced version of point inserted in the Hnsw structure. It contains original id of point as submitted to the struct Hnsw an ordered (by distance) list of neighbours to the point and it position in layers.