hyperstreamdb::core::index::hnsw_rs

Module libext

This file contains lib to call hnsw from julia (or any language providing a C api) The AnnT trait is implemented with macros for u32, u16, u8, f32, f64 and i32.
The macro declare_myapi_type! produces struct HnswApif32 and so on.

Structs

DescriptionFFI

This structure provides a light description of the graph to be passed to C compatible languages.

HnswApiNodata

HnswApif32

HnswApif64

HnswApii32

HnswApiu8

HnswApiu16

HnswApiu32

Neighbour_api

The basic Neighbour info returned by api

Neighbour_api_parsearch_answer

Neighbourhood_api

The response to a neighbour search requests

Vec_api

to be able to return a vector from rust in a julia struct before converting to a julia Vector

Functions

drop_hnsw_f32⚠

drop_hnsw_u16⚠

file_dump_f32

file_dump_i32

file_dump_u8

file_dump_u16

file_dump_u32

init_hnsw_f32

init_hnsw_i32

init_hnsw_ptrdist_f32

init_hnsw_ptrdist_i32

init_hnsw_ptrdist_u8

init_hnsw_ptrdist_u16

init_hnsw_ptrdist_u32

init_hnsw_u8

init_hnsw_u16

init_hnsw_u32

init_rust_log

to initialize rust logging from Julia

insert_f32

insert_i32

insert_u8

insert_u16

insert_u32

load_hnsw_description

returns a const pointer to a DescriptionFFI from a dump file, given filename length and pointer (*const u8)

load_hnswdump_NoData_DistNoDist

load_hnswdump_f32_DistCosine

load_hnswdump_f32_DistDot

load_hnswdump_f32_DistJeffreys

load_hnswdump_f32_DistJensenShannon

load_hnswdump_f32_DistL1

load_hnswdump_f32_DistL2

load_hnswdump_i32_DistHamming

load_hnswdump_i32_DistL1

load_hnswdump_i32_DistL2

load_hnswdump_u8_DistHamming

load_hnswdump_u8_DistJaccard

load_hnswdump_u8_DistL1

load_hnswdump_u8_DistL2

load_hnswdump_u16_DistHamming

load_hnswdump_u16_DistL1

load_hnswdump_u16_DistL2

load_hnswdump_u16_DistLevenshtein

load_hnswdump_u32_DistHamming

load_hnswdump_u32_DistJaccard

load_hnswdump_u32_DistL1

load_hnswdump_u32_DistL2

new_hnsw_f32

same as max_layer with different arguments, we pass max_elements and max_layer

new_hnsw_u16

parallel_insert_f32

parallel_insert_i32

parallel_insert_u8

parallel_insert_u16

parallel_insert_u32

parallel_search_neighbours_f32

search nb_vec of size vec_len. The the searches will be done in // as far as possible.

parallel_search_neighbours_i32

search nb_vec of size vec_len. The the searches will be done in // as far as possible.

parallel_search_neighbours_u8

search nb_vec of size vec_len. The the searches will be done in // as far as possible.

parallel_search_neighbours_u16

search nb_vec of size vec_len. The the searches will be done in // as far as possible.

parallel_search_neighbours_u32

search nb_vec of size vec_len. The the searches will be done in // as far as possible.

search_neighbours_f32

search_neighbours_i32

search_neighbours_u8

search_neighbours_u16

search_neighbours_u32