hyperstreamdb

Module python_distance

Structs

PySparseVector

Sparse vector representation storing only non-zero indices and values

Functions

py_cosine

Compute cosine distance between two vectors

py_cosine_batch

Compute cosine distances between a query vector and multiple database vectors

py_cosine_sparse

Compute cosine distance between two sparse vectors

py_hamming

Compute Hamming distance between two vectors

py_hamming_auto

Compute Hamming distance with auto-packing support for unpacked binary vectors

py_hamming_batch

Compute Hamming distances between a query vector and multiple database vectors

py_hamming_packed

Compute Hamming distance between two bit-packed binary vectors

py_inner_product

Compute inner product (dot product) between two vectors

py_inner_product_batch

Compute inner product between a query vector and multiple database vectors

py_inner_product_sparse

Compute inner product between two sparse vectors

py_jaccard

Compute Jaccard distance between two vectors

py_jaccard_auto

Compute Jaccard distance with auto-packing support for unpacked binary vectors

py_jaccard_batch

Compute Jaccard distances between a query vector and multiple database vectors

py_jaccard_packed

Compute Jaccard distance between two bit-packed binary vectors

py_l1

Compute L1 (Manhattan) distance between two vectors

py_l2

Compute L2 (Euclidean) distance between two vectors

py_l1_batch

Compute L1 distances between a query vector and multiple database vectors

py_l2_batch

Compute L2 distances between a query vector and multiple database vectors

py_l2_sparse

Compute L2 distance between two sparse vectors